Implement a better friend-highlighting algorithm
This commit changes the friend-highlighting algorithm
from a naive approach that resolved conflicts by choosing
the better friend to a more sophisticated one that uses
the historical data of which friends you've seen with
which other friends, using those trends to guess the most
likely friend, with the better friend being chosen in the
event of a tie.

This commit also updates the README.md with more commands
and clearer organization and descriptions.