Make activities on same day chronologically ordered
Previously, two activities listed on the same day would be saved to
the friends.md file in the reverse order of their writing, which
anecdotally is not the order in which they make the most sense.

This commit changes the behavior, so if one adds the activities:
"2016-01-01: Got breakfast."
"2016-01-01: Got lunch."

The lunch activity will appear as "later" than the breakfast one
when listing activities.

This resolves #37.