Restructure code and add integration tests
This commit restructures the `bin/friends` code
to be more modular and call commands that are
outside of the `bin/` directory.

In addition, it changes all tests from unit tests
to integration tests. These integration tests
that run the entire CLI program in a process are
easier and more accurate than unit tests which
in this case required lots of stubbing to test
basic behaviors, and could not easily test things
like STDOUT or exit statuses.

This commit also makes a small change to the
README to reflect the nature of this repository
now firmly being a CLI and not a general-purpose
library.

Lastly, this commit makes some small tweaks to
general configuration, test setup, etc. to
make it generally more robust and modern.

Closes #127.