Add --since and --until flags to `graph` and `list activities`
This commit adds optional `--since` and `--until` flags that
can be used to filter down the activities displayed in the
`graph` and `list activities` commands.

It also changes the behavior of `graph` so that leading and
trailing months with no activities are not graphed.

Lastly, this commit fixes a typo in the README that was causing
a table-of-contents link to not work.

Closes #153.