Separate optional test dependencies with Gemfiles
This commit adds a `Gemfile.old` which is used for testing
old Ruby versions in Travis. This file does not include
dependencies such as Rubocop and SimpleCov that are only
used in either local development or our Travis build for
the latest Ruby version.

Included in these changes is pinning the SimpleCov dependency
to the version fixed by
https://github.com/colszowka/simplecov/pull/747 to fix tests.

[@Shen-Sat](https://github.com/shen-sat) was instrumental in
discussing these changes and coming up with many parts of
the overall approach. Many thanks!

Fixes #238